feat: add native xml support
This should make it easier to parse the xml file and get more
information out of it.

This is done by adding a new git submodule for https://github.com/Phrogz/SLAXML
and adds a few basic tests to show case how it works.

These test files are executed using `busted`.

Relates-To: https://github.com/NightQuest/comicmeta.koplugin/issues/1

Change-Id: 0de6160bdc646e5c54589d46172bb0a9